+++ src/components/AvatarGroup/AvatarGroup.stories.tsx
@@ -7,7 +7,7 @@ import { ConnectTheme } from '../ConnectTheme';
 import { GradeBand } from '../../enum/gradeband';
 
 const meta: Meta<typeof AvatarGroup> = {
-  title: 'WIP/AvatarGroup',
+  title: 'Misc/AvatarGroup',
   component: AvatarGroup,
   tags: ['autodocs'],
   parameters: {